pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Kieran Brooks", "James Giblin", "Nikhil Salvi", "Maalav Kanekar", "Austin Syban", "Sajad"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pep517 = "0.13.1"
speechrecognition = "^3.10.1"
PyAudio = "0.2.14"
Pillow = "10.2.0"
tk = ""
pyttsx3 = "^2.90"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"